card update

+++ src/main/java/org/anonymous/card/controllers/TrainController.java
@@ -28,7 +28,7 @@ public class TrainController {
     }
 
     @GetMapping("/predict")
-    public List<Integer> predict(@RequestParam("data") String data) {
+    public List<Long> predict(@RequestParam("data") String data) {
         List<Integer> items = Arrays.stream(data.split("_")).map(Integer::valueOf).toList();
         System.out.println("items:" + items);
         return predictService.predict(items);
